# app/album/entities.go
package album

type Album struct {
	ID     uint    `json:"id" gorm:"primary_key"`
	Title  string  `json:"title"`
	Artist string  `json:"artist"`
	Price  float64 `json:"price"`
}

type CreateAlbumInput struct {
	Title  string   `json:"title" binding:"required"`
	Artist string   `json:"artist" binding:"required"`
	Price  *float64 `json:"price" binding:"required"`
}

type DeleteAlbumInput struct {
	ID uint `json:"id"`
}